requirements.txt
modelcontextprotocol
mcp[cli]
langdetect
langid
streamlit
openai
python-dotenv
ipykernel
httpx
pymongo
tiktoken

